svn path=/branches/work/koffice-isi/; revision=891282

+++ filters/kspread/excel/sidewinder/format.h
@@ -85,7 +85,7 @@ public:
 
   unsigned style;
 
-  unsigned width;
+  float width;
 
   Color color;
